Refactor LocalProvider into mixin modules
